refactor redux state
