--- src/components/Random/Random.module.css
@@ -1,38 +1,0 @@
-.counter {
-  border: 1px solid lightgray;
-  margin: 36px 24px;
-  padding: 24px;
-  width: 240px;
-}
-
-.header {
-  font-size: 24px;
-  font-weight: normal;
-  margin: 0 0 12px;
-  text-align: center;
-}
-
-.button {
-  background: lightseagreen;
-  border: none;
-  border-radius: 5px;
-  color: white;
-  cursor: pointer;
-  display: block;
-  font-size: 16px;
-  margin: 0 auto 24px;
-  padding: 12px 24px;
-  text-shadow: 1px 1px 1px rgba(0, 0, 0, 0.5);
-  white-space: nowrap;
-}
-
-.button:disabled {
-  background: lightgray;
-  cursor: not-allowed;
-}
-
-.button:active:not(:disabled) {
-  left: 1px;
-  position: relative;
-  top: 1px;
-}
