--- src/components/Random/index.js
@@ -1,1 +1,0 @@
-export {default} from './Random';
