--- src/features/counter/actionTypes.js
@@ -1,2 +1,0 @@
-export const INCREMENT_COUNTER = 'INCREMENT_COUNTER';
-export const DECREMENT_COUNTER = 'DECREMENT_COUNTER';
