--- src/features/openedModals/actionTypes.js
@@ -1,4 +1,0 @@
-export const OPEN_DELETE_MODAL = 'OPEN_DELETE_MODAL';
-export const OPEN_EDIT_MODAL = 'OPEN_EDIT_MODAL';
-export const OPEN_NEW_ENTRY_MODAL = 'OPEN_NEW_ENTRY_MODAL';
-export const CLOSE_MODAL = 'CLOSE_MODAL';
