--- src/features/openedModals/index.js
@@ -1,7 +1,0 @@
-export {default as openedModals} from './openedModalsReducer';
-export {
-  openDeleteModal,
-  openEditModal,
-  openNewEntryModal,
-  closeModal,
-} from './actionCreators';
