--- src/features/pagination/actionTypes.js
@@ -1,5 +1,0 @@
-export const RESET_PAGE = 'RESET_PAGE';
-export const SET_PAGE = 'SET_PAGE';
-export const INCREMENT_PAGE = 'INCREMENT_PAGE';
-export const DECREMENT_PAGE = 'DECREMENT_PAGE';
-export const SET_ROWS_PER_PAGE = 'SET_ROWS_PER_PAGE';
