--- src/features/pagination/index.js
@@ -1,8 +1,0 @@
-export {default as pagination} from './paginationReducer';
-export {
-  resetPage,
-  setPage,
-  nextPage,
-  prevPage,
-  setRowsPerPage,
-} from './actionCreators';
