--- src/features/random/actionTypes.js
@@ -1,1 +1,0 @@
-export const GET_RANDOM_NUMBER = 'GET_RANDOM_NUMBER';
