--- src/features/random/index.js
@@ -1,3 +1,0 @@
-export {default as RandomReducer} from './RandomReducer';
-export {default as useRandomAPI} from './selectors';
-export {default as useActions} from './actionCreators';
