--- src/features/rowId/actionCreators.js
@@ -1,6 +1,0 @@
-import {SET_ROW_ID} from './actionTypes';
-
-export const setRowId = id => ({
-  type: SET_ROW_ID,
-  payload: id,
-});
