--- src/features/rowId/actionTypes.js
@@ -1,1 +1,0 @@
-export const SET_ROW_ID = 'SET_ROW_ID';
