--- src/features/rowId/index.js
@@ -1,2 +1,0 @@
-export {default as rowId} from './rowIdReducer';
-export {setRowId} from './actionCreators';
