+++ src/features/studentsData/index.js
@@ -1,0 +1,17 @@
+export {default as studentsData} from './tableDataReducer';
+export {
+  deleteEntry,
+  addEntry,
+  editEntry,
+  sortEntries,
+  setRowId,
+  resetPage,
+  setPage,
+  nextPage,
+  prevPage,
+  setRowsPerPage,
+  openDeleteModal,
+  openEditModal,
+  openNewEntryModal,
+  closeModal,
+} from './actionCreators';
