--- src/features/tableData/actionTypes.js
@@ -1,4 +1,0 @@
-export const DELETE_ENTRY = 'DELETE_ENTRY';
-export const ADD_ENTRY = 'ADD_ENTRY';
-export const EDIT_ENTRY = 'EDIT_ENTRY';
-export const SORT = 'SORT';
