--- src/features/tableData/index.js
@@ -1,2 +1,0 @@
-export {default as tableData} from './tableDataReducer';
-export {deleteEntry, addEntry, editEntry, sortEntries} from './actionCreators';
